Improved responsiveness: File frames load in array when file is opened
Old strategy was to find the last frame without loading the previous ones. This is usually impossible, as frame decoding depends on previous frames, so imageio does this automatically. When changing the selected frame, finding a given frame on the file depends on finding reading again all previous frames.

Now, all frames are decoded and stored in MediaFIle_t.iioArray, so slidebar responsiveness is hugely improved